Update: searching the fishname
I changed a method of searching for fish:
Before: I used a 'for loop' to search the fish with input fishname, it worked. But if I added a condition (else{res.render('fishnotfound'), only the first item in the loop is working (if i search other item in the loop, the page will be directed to 'fishnotfound'). I assum the searching ends after the 1st iteration??!!!
Now: instead of for loop, I use findIndex() to look for the location of the input in an array. And then I assign the page to show the associted information. And it works well.

To make the searching easier, I also changed the user input and the fish names into uppercases before searching. By doing so, the information can be searched in both upper and lower cases